General update
Clean up in scripts.
New folders added for better organization.
Textures for GUI added.
Textures for prefabs boosters
Added Scripts for all boosters.
Ammunition stock implimented
Grenades implimented
Defence/Health and Damage have been added

+++ New Unity Game/Assets/scripts/SpawnController.cs
@@ -6,12 +6,10 @@ public class SpawnController : MonoBehaviour
 	
 
 	public GameObject Skull;
-	//GameObject enemySpawn2;
-	//GameObject enemySpawn3;
-	
+		
 	public float spawnTimer;
 	public float howFast;
-	public short maxEnemies;
+	public int maxEnemies;
 
 
 	void Start ()
